pyproject.toml
[tool.poetry]
name = "arachne"
version = "0.1.0"
description = ""
authors = ["suriya <suriyaganesh097@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-playwright = "^0.5.1"
pydantic = "^2.8.2"
structlog = "^24.4.0"
fastapi = "^0.112.2"
boto3 = "^1.35.10"
regex = "^2024.7.24"
jupyterlab = "^4.2.5"
langchain = "^0.2.16"
langchain-community = "^0.2.16"
langchain-core = "^0.2.38"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
